+++ activerecord/CHANGELOG.md
@@ -18,4 +18,8 @@
 
     *Erol Fornoles*
 
+*   Adds calculations on multiple columns
+ 
+    *Jacob Kjeldahl* 
+    
 Please check [5-0-stable](https://github.com/rails/rails/blob/5-0-stable/activerecord/CHANGELOG.md) for previous changes.